Make QRCode canvas reuseable
